fix bug preventing awobmolg.alda from eval-ing at the CLI
The Alda score evaluator was throwing an error about not being able to resolve the symbol $current-offset, when trying to do "boot parse --map -f test/examples/awobmolg.alda". Interestingly, I was still able to evaluate the score from the REPL within the alda.lisp namespace. The problem ended up being in alda.lisp/chord*, a macro that was generating a literal symbol '$current-offset, which was only eval-able within the alda.lisp namespace. Fully qualifying the symbol (`$current-offset) fixed it. :+1: