fix part creation & corresponding tests
Aliasing instrument names (e.g. piano, trumpet) to MIDI instruments (e.g. midi-acoustic-grand-piano, midi-trumpet) illuminated a bug where existing instruments weren't being recognized because it was looking at the stock instrument rather than the actual "name" of the instrument. For example, calling a part called "piano" would initialize a midi-acoustic-grand-piano, but then subsequently calling "piano" again would look for something with the stock instrument "piano," not find it, and decide to initialize a new midi-acoustic-grand-piano instead. The "name" of the instrument is stored in its id, so now the part-assigning function looks at the beginning of the id string (+ a hyphen) to determine if an instrument already exists. Tests pass now. :+1: