alda.repl is fast now! :rocket:
It turns out that the slowness was just the result of opening a new MIDI synth every time a new score is evaluated. The parsing process actually happens almost instantaneously. So now, the Alda REPL opens one MIDI synth initially and uses it to play the score generated by each line.

One-off playing of Alda code/files from the command-line is still pretty slow, but that's probably a JVM startup time issue.