always use a new MIDI channel for each instrument
Initially I was trying to save channels (since there are only 16) by doubling up on channels when there are two instances of the same MIDI instrument.

However, while transcribing the prelude to Bach's first cello suite I ended up needing to create a second cello instance (which was made to share a MIDI channel with the first one) for the overlapping 'A' pedal tone notes alternating between the open A string and unison 'A' note on the D string, and I discovered that the MIDI note-off events for that 'A' were clobbering those of the other part.

Having two instances of an instrument share a MIDI channel is clearly asking for trouble. From now on, one MIDI instrument (instance) = one MIDI channel.