R.I.P. .idea folder
As more people are starting to become involved with the development in Alda, I think it will be easier to collaborate and discuss feature ideas, etc. through some combination of the wiki and the issue tracker. I've moved the ideas in the .idea folder to the wiki, and moved the midi-patch-demo script to the bin folder.