refactor, hammer out kinks
Cleaned up REPL commands -- pull out useful information like the parsing context and the REPL ConsoleReader into dynamic vars in a new shared namespace, alda.repl.core. Include new :play command in alda.repl.commands.

Use alda.now/refresh! to reset the state of the score itself (not the score text) so that all the instruments start at the same time (0) and we are not replaying the score onto itself, so to speak.

Bugfix: When playing parsed code via `alda.repl.core/interpret!`, if the parsing context is `:score`, skip the `score` part and use the current score-in-progress instead -- otherwise, it would reset the state of the score-text to "" and wipe clean all the events except for the new ones that you're playing.