bugfix: voice group + inst call = parse failure
I just noticed that following a voice group with a new instrument call causes a parse failure because the "voices" rule is written in such a way that it must be followed by either "V0:" or the end of the file.

This fixes that by adding a lookahead to a new part as another option to that rule.