bugfix: make playback start immediately
There is a bug here that causes playback to always start at offset 0, rather than the earliest offset out of the event set. I think the intent here is to make sure that the offset is at least 0, but we still want it to be the earliest offset from the event set, as long as it's a positive number (which it should be anyway... but it doesn't hurt to have a safety net!).