Extract parser into separate tasks :scissors:
This commit changes Alda's parsing process from using a single, monolithic grammar to parse an entire score in one pass, to parsing a score in several discrete steps, using a separate grammar/parser for each phase.

The perks of doing it this way:

- It will give us greater flexibility with syntax features. For example, this will make it a lot easier (maybe even possible to begin with) to implement inline variable definitions which terminate in a newline -- we can just parse them after comments and Clojure expressions, but before everything else.
- We can leverage more of Instaparse's built-in optimizations for parallelizing work. Each instrument part will be parsed concurrently, for example.
- Because we'll be feeding each parser smaller bodies of text, performance will improve significantly. I tested at least one score that took ~6 seconds to parse before, but now takes ~600-700 ms.

The downsides:

- There is no longer a single parse tree for an entire score.
- Parsing errors will be harder to understand from the stacktrace alone. Line/column numbers will no longer refer to the entire score file, but rather the small section of the score that is being parsed.
- We'll need to rewrite/rethink `parse-with-context`, which the Alda REPL uses to determine the context of a snippet of Alda code. Currently this works by doing a series of partial parses using the single alda.bnf grammar.

I think these are hardships worth enduring for the benefits we'll gain, especially in light of the poor performance we're dealing with with the current single-grammar system, and the need to pave the way forward for advanced syntactical features as the language's grammar grows in complexity.