update tests
Previously, `test-parse` was used for all the parser tests to attempt to parse each syntax example as that rule, e.g. `(test-parse :cram "{c c c}")`. The implementation relied on our previous single-grammar parser, parsing from a given start rule. This is no longer possible, but we do have a general-purpose `parse-with-context` function which can attempt to parse arbitrary input as music-data. This is actually better for tests because we're not telling the parser which
rule to use -- we're just handing it some code and saying "parse this as music data," and testing that it parses as the right thing.