server: rm play, parse and script tasks
With a dedicated client, we'll be moving away from having a CLI interface for the server at all. All interaction will be client -> server. Will probably end up moving the "repl" and "server" entrypoints into their own namespaces, and have the client call them directly, then scrap the alda.cli namespace altogether.

(The "repl" will eventually be ported to the client and just be another way to interact with a server.)