load FluidR3 when `stock` option not included
This also includes a change to the way we're holding a reference to the FluidR3 soundfont. Trying to get the soundbank from the same InputStream more than once results in an error "cannot get soundbank from stream." The workaround here is to only get the soundbank once, then store that in *midi-soundbank*. The `load-fluid-r3!` function checks to see if *midi-soundbank* is already loaded (non-nil), and iff it is not, it fetches FluidR3 and loads it into *midi-soundbank*.

This change makes it so that you can start the server and/or repl multiple times in a Clojure REPL while developing.