don't write target dir (not needed in this case)
The `bin` and `exe` tasks write binary executables to an output dir specified as an option to the `build` task, and this happens separately from writing to the target dir, so we don't really need to use the `target` task here, which might save a little time during builds.

When we do want the target dir written to, we can just add `target` on the end of the task we're running on the command-line.