Build uberjar with JDK 1.7 support.
`boot build` will now build uberjars (and therefore binaries) that are compatible with JDK 1.7+, making it so that people with older Java versions can use Alda.

I tried going back to JDK 1.6, but there are a couple incompatible things in the Java client source that would need to be resolved, and I'm not sure if it's worth the effort at this point, with JDK 1.6 on its way out. We can always revisit in the future if there is a need for Java 6 compatibility.

Blockers for Java 6 compatibility:

1) The switch statement in Client.java operates on strings, which is not supported in Java 6. We could just make it a series of if/else if statements.

2) The use of `.inheritIO()` in Util.java here:

   public static void runProgramInFg(String... args)
   throws IOException, InterruptedException {
     new ProcessBuilder(args).inheritIO().start().waitFor();
   }

   I'm sure there is some pre-JDK7 workaround for this.