bugfix re: voice parsing
In some cases, voice parsing wasn't happening correctly -- subsequent voices were being nested inside of the previous voice, instead of them all being nested at the same level in the voice group like they should be. This was caused by an ambiguous grammar rule that allowed a voice group to be an event in a voice. Alda doesn't support nesting voices inside of voices, so we can solve the ambiguity by disallowing that explicitly.