only set timbre level in alda.lisp
This is the only real place we need to call alda.util/set-timbre-level!, which defaults the logging level to :warn if it isn't already set in the environment.

Typically things that use Alda (and thus may include logging) will require alda.lisp, so the logging level will get set at that point.