updated alda.now to use alda.lisp the new way
With the changes I just made to alda.lisp, score evaluation is now a self-contained process. This allows us to be flexible in the way we use alda.now.

I've adjusted things so that the `play!` function by default will start up a new score and evaluate the events in its body within that context. It actually uses *current-score* as the context, but when that's nil, it starts a new score and uses that.

Scores are represented as atoms referencing Alda score maps.

*current-score* can be manipulated via the `with-score` and `with-new-score` macros, which bind *current-score* to a particular score or a new score, respectively, and then both modify *current-score* and return the modified score.

There's also `play-score!`, which will play an entire score, given the score map or an atom referencing it.