adjust alda.sound to handle isolated scores
This changes all of the alda.sound functions to accept an "audio context" atom, which references a Clojure map whose values include mutable things that do the work needed to bring music to your eardrums -- things like the MIDI Synthesizer and the JSyn SynthesisEngine. The audio context also includes score-specific information like which audio types have been initialized, and the mapping of instrument IDs to MIDI channels in the MIDI synth being used for that score.

Luckily, in testing this locally, there don't seem to be any obvious performance issues so far -- it appears that multiple MIDI synthesizers and multiple SynthesisEngines can coexist peacefully -- which is awesome. We'll see if this causes problems in the long run.

One issue is that it takes a second or two for a new MIDI synthesizer to initialize. Whereas before, this was a one-time cost paid when alda.sound was evaluated, now it happens every time a new score is created and the MIDI audio type is set up. There is a mechanism for setting it up ahead of time via the `set-up!` function, but it's still not great that you have to wait a second or two every time you start a new score. I have an idea for fixing that that I think will work -- set up a pool of MIDI synths that have already been initialized in background threads, and draw from the pool each time a fresh synth is needed. There could perhaps be a supervisor thread that monitors the number of available MIDI synths and spins up new ones as needed. It would be good to handle stopping the MIDI synths after we don't need them any more... not sure how to approach that just yet. Maybe just provide a way to do it on demand (I guess there already is a way -- the tear-down! function) and have the REPL and server do it whenever you start a new score or load an existing one.