bugfix: don't let attributes be applied more than once
Now attributes are only applied right before a note or rest event happens. Before this commit, there was one other place that attributes were applied -- at the end of the score. I was initially thinking that would be good so that the current "state" of the instrument at the end of the score reflects any attribute changes that may happen at the very end of the score. However, the attributes would then be applied again once the score is continued, leading to problems for attributes whose values
depend on the current attribute value (e.g. octave up/down).