don't let durations end with a slur
This prevents ambiguity where either a note's last duration ends with a slur or the note itself ends with a slur.

After thinking about it, I realized that a duration (at least one that can be considered "slurred") never exists in an Alda score unless it's associated with a note, so we don't really need durations to include a slur. (Other things can have durations too, like cram expressions, but those can't be slurred, so it's probably for the best that we not allow them to be slurred.)