simplify variable getting/setting
This simplification makes variable getting/setting a whole lot easier to understand and reason about. Instead of tracking the environment of a variable at the time it is defined and then referencing it later, we just look through the variable's events and go ahead and dereference any get-variable events that might be in there. This ensures that no defined variable will contain a "get-variable" event. Scope is no longer an issue -- we just use the current scope at the time the variable is defined.

With this change, things basically work exactly the same as before, but now the Alda score map contains a single key `:env` instead of a complicated, nested `:variables` map and a temporary `:env` to track the current scope.

I suspect that if we ever need to make this more complicated (e.g. first class functions/closures), we can leverage Clojure functions instead of reimplementing scope... it seems like we don't need to reimplement scope as part of the Alda score evaluation state machine.