make server stateless (remove stateful commands)
Prior to this commit, the server was responsible for not only handling requests, but also holding a score in memory and appending/replacing it or providing info about it based on the clients' commands.

Commands based on this state included:

- alda play (without a file or string of code -- plays the score in memory)
- alda play --append
- alda score
- alda info

This commit removes these commands (and associated code) from the client and removes the stateful parts of the server, so that the server only has to worry about starting an Alda environment (loading MIDI synth pool) and responding to one-off requests to play or parse a score.

The next step will be to offload the "work" (having an open MIDI synth, parsing and playing scores) onto separate worker processes and having the server just be a broker between the client and the workers.

With this type of architecture, it is not really feasible for the server or the workers to keep track of a "current score" for the client, so the user will have to do this manually. I think this is actually a better pattern, because people who use computers are already used to managing documents as files and running command-line programs as one-off things that don't manage state for us.

I think my original intent was for the Alda client/server experience to be just like the Alda REPL experience, which was an interesting experiment, but unnecessary I think. If users want the REPL experience, they can always use the REPL.