generate an ID for each player process, include it in logs
I considered doing something more sophisticated that would be more resistant to
collisions (e.g. UUIDs), but after thinking about it, I concluded that our needs
in that area are really basic and it would be better to make IDs easier for a
user to type.  At some point, Alda will allow users to start a score, and then
continue to interact with the player process by specifying which player process
to use.

Another option is to use the port as an ID, but I think it's easier to
differentiate between letters than numbers.

Collisions are unlikely because 26 * 26 * 26 = 17,576 possible values, and there
will only be a handful of player processes running on a single machine at a
time. It doesn't matter if an ID is recycled later, because these player
processes are ephemeral and we just need a way to differentiate between the
handful that are running at any given time.