alda play: exit immediately after sending instructions to the player
I had this set up so that in the case where you're spawning a player to play a
score, the process would stay open until you press ^C, and at that point, we
would trap the SIGTERM and send a shutdown message to the player.

This is no longer necessary because player processes will now shut themselves
down after a period of inactivity.