player: add (currently unused) `/system/offset` address
I was thinking that I might need this for the Alda v2 REPL `:play` command, but
then after thinking/coding for a while, I realized that it would be
easier/better to shutdown the player process, pick up a new one, and send the
entire score again! Sometimes a sledgehammer is the right tool for the job. :)