Beats() signature change: (float64, error) => float64
Notionally, I can't think of a reason why Beats() should ever return an error. I
think the only reason why I included an error argument was so that I could
return a clear non-value in the case of miliseconds => beats conversion. But
milliseconds => beats conversion _is_ possible, it's just that I've been punting
on it (and am continuing to punt now). Eventually, if I want to implement
Beats() for NoteLengthMs, I'll need to adjust the signature again to include a
tempo argument (i.e. `Beats(tempo float64)`). Until then, I think it's OK to
panic in that scenario. In practice, it shouldn't be possible for the end user
to go down that code path.