hopefully fix the `find` failures for good
I think the issue is that if the `target` directory as a whole hasn't been
touched in over 7 days, then we delete the `target` directory, and then we try
to delete other directories inside of `target`, but now that fails because they
don't exist anymore, because we deleted the whole `target` directory.

The addition of `-not -path target/` will exclude `target/` from the `find`
results, which is an improvement. I'm like 80% sure this will fix the issue. If
it doesn't, it'll be back to the drawing board.