windows build experiment: don't do the initial `alda version` (see message)
As with most commands, `alda version` secretly starts player processes in the
background. I'm wondering if maybe the issue here is related to the subsequent
`alda doctor` process trying to talk to one of those player processes, which
might not work in CI and just on Windows for some reason. This might have
something to do with the fact that we didn't start the player processes in "no
audio" mode, or it might be something Windows-specific.