Implement a (hopefully!) robust `alda update` mechanism
I haven't yet implemented the part where use the Alda API to check for versions
newer than the one installed. This commit is just the actual update mechanism.
`alda update` takes a flag that lets you specify which version you want to
install. That works, as of this commit, and hopefully I've done it in a way that
will work on Windows, macOS, and Linux. CI will tell me whether or not this is
the case when I push this commit.