Improvements to handling situations where alda-player isn't installed or is the wrong version
In either of these scenarios, `alda doctor` will now offer to install the
correct alda-player version.

I've updated the relevant user-facing error messages to recommend running `alda
doctor` to remedy the situation.

This incidentally also improves the way that we've been spawning player
processes. I realized that for _each_ of the player processes that we spawn (and
it can be up to 3 at once, I think?), we first invoke `alda-player info` to
check the `alda-player` version against the `alda` version, but we really only
need to do this once.