Experiment: explicitly preserve os.Environ() when spawning a process
According to https://pkg.go.dev/os/exec#Cmd (see "Env" in the Cmd struct):

  If Env is nil, the new process uses the current process's environment.

Based on that, it seems like it should _already_ be the case that we are
preserving important environment variables like XDG_CACHE_HOME when we
spawn player processes. But a user's report on Slack really makes it
seem like XDG_CACHE_HOME isn't being preserved and the player process
is using the default cache directory instead, so that there is a
mismatch between the client and the player.

I'm trying this as an experiment to see if maybe the exec.Command
behavior is different on different OS's. Maybe on some OS's, the quote
above is a lie, and it isn't really preserving environment variables?