Temporarily checking in Queue class
I implemented this simple Queue class in order to have something like
LinkedBlockingQueue in JS. It works great, but it doesn't support blocking
receive, since JS is single-threaded.

After looking into this for a while, it seems like Kotlin's coroutines and
channels solve the same problem, so I'm giving that a try now. Checking this in
so that I have it in git history in case I ever need it in the future.