Address warnings about synchronizing on Booleans by just not synchronizing ¯\_(ツ)_/¯
I'm not convinced that we ever really needed to. I think I was just coding
defensively. But my understanding of the issue here is that when you synchronize
on a boolean, you aren't really synchronizing because every instance of true or
false is equivalent. So it wasn't really synchronizing before anyway, so
removing the `synchronized(...) { ... }` shouldn't be any different from the
behavior before.