We'll need to use version 2.2.6+ now for the `alda update` test (see message)
We've been using version 2.0.0 with no issues up until now, but now CircleCI is
getting rid of the older macOS runners and we have to start using an M1 runner.
But there's no darwin/arm64 build for version 2.0.0. In fact, there's no
darwin/arm64 build for _any_ version yet. But there will be, starting with
version 2.2.6.

After releasing version 2.2.6, I will un-comment these and it should all be
working again. But I have to comment them out for now in order to get the build
passing so that I can release version 2.2.6.