`announce-release`: link to release notes instead of including the Markdown in the Slack message
The Markdown wasn't ideal because it often wouldn't format quite right, and
reference links [like this] where the "[like this]: https://link.here" part
wasn't included because it's at the bottom of the changelog, resulted in the
links being missing.

I also just ran into an issue where the changelog text was longer than Slack
allows, so it wouldn't post it at all.

So I decided to just link to the release notes on GitHub.