feat: Add --wait flag to 'alda play'
This flag allows the user to block the 'alda play' command until playback is complete.

This is implemented by adding a new "finished" state to the player process. The client sends a '/system/playback-finished' message at the end of the score, which transitions the player to the "finished" state. The client then polls for this state.

This change is implemented in a way that does not affect the live coding use case, as the new messages and states are only used by the 'alda play' command.

This addresses issue #514.